Added QUAL to list of columns to remove prior to consensus calling
Added QUAL to list of columns to remove before calling the consensus. Vardict actually generates values for this column which causes it to be set on it's own line in the data. In paired calling QUAL is added to the somatic calls, so it is in the FORMAT column. But in the tumour only calling this isn't in FORMAT